Use taggedtuple instead of namedtuple when defining ShardingSpecs
Because namedtuples don't take the class into account when comparing for
equality!
